Started implementing multi-slot savegames. To be compatible with DOS 8.3 names and VFAT case-insensitive names, files will be created with <slot#>.iff names. The location of the files is defined as follows:
- In UNIX savegames will be stored in $HOME/.sarien/<gameid>/.
  HOME is always defined.

- In DOS, Windows 9x/NT/2000 they'll be created in
  %USERPROFILE%\Sarien\<gameid>\. If USERPROFILE is not defined,
  it will use \Sarien\<gameid>\ (which seems to be a reasonable
  fallback in DOS and Windows 9x systems).

In both cases, if the game ID is not defined, the file will be
stored in the same directory level as the <gameid> subdirectory.
Every game should set an ID. If it's not set, perhaps Sarien
should assign them an ID based on CRC.

This implementation uses the mkdir(2) call, wich is SVr4, POSIX,
BSD, SYSV and X/OPEN compliant. I hope it's also provided by
the Watcom libc and Win32 APIs.