Merged Ryan Gordon's SDL patches:
1) configure.in is patched to include the correct linker/compiler flags
for building SDL-related code. By default, it can forget to link in the
pthreads library, etc. The use of the sdl-config script handles this, now.

2) Programs should include "SDL.h", not <SDL/SDL.h> for portability
reasons. This is fixed in the sound and graphics drivers (and again,
sdl-config makes sure this form of inclusion works).

3) PGUP, PGDOWN, HOME, and END now work like the 7, 9, 1, and 3 numpad
keys, respectively, so I can make Roger Wilco walk on diagonals on my
keyboard.  :)