Fixed DICE build for the Amiga port. From the DICE docs,
Pragma argument count conflict with prototype: %.*s
---------------------------------------------------
    This error indicates that there is a difference of opinion as to the
    number of parameters that an inline function takes.  The #pragma
    statement for the function codes the number of parameters while the
    prototype explicitely lists them.  If you encounter this error, you
    will have to examine the two in order to determine which is right.

The DICE port uses the pragmas instead of the prototypes. Also filling
ximage with palette[0] (black) to clear the screen before starting,
fixed duplicate "Sarien" word in the title bar and adjusted bottom/
rightmost row of pixels in put_block to correctly restore the screen
when scale > 1. Set second parameter (struct timezone) of gettimeofday()
to NULL; OS includes didn't seem to define it. Sarien.DICE added to
the list of distribution files in the toplevel Makefile.