From: Thomas Akesson <tapilot@home.se>
Currently the word list is expanded in memory and an agi_word struct is
created for each word. I developed an algorithm taking advantage of the
Sierra words.tok format which very cleverly combines compression and fast
searches. I have done some basic testing and it seems to work fine but I
have to check the memory management again.

Memory:
A typical words.tok is 6-7 kB with some 1000-1100 words. This expands to
typically 1 kB more of string data (including nulls but excluding ids). The
agi_word struct is probably 8 byte which adds another 8 kB. The "Sierra
style" algorithm uses the original words.tok format which saves some 9 kB
of memory.

This is of course peanuts on most modern system. The major issue is that
words.tok can stay in storage RAM on PalmOS which saves 17 kB of dynamic RAM.

Speed:
I timed the two algorithms when finding each word in a list 100 times
(approx 100000 find operations). The old algoritm takes more than 6 seconds
(PIII 550) while the "Sierra style" algorithm takes 0.55 seconds.

The reason for the huge difference is that the current find_word uses a
linear algorithm (not Divide and Conquer as claimed by the comment). Using
a real Divide and Conquer algorithm, like binary find, might be slightly
faster than the "Sierra style" but would use the same amount of memory as
the current find_word.